[feat] refactored PoupUpTaskList #32 (#33)
merged
